Optimise formatting
This commit gets rid of a violation of the DRY principle. In the previous version of the section, it was possible to miss the link to the relevant documentation and the information, that there are variants was repeated twice. Now, in the new version, this is solved and the information is displayed just once.