Add local sources first
I use an Electron-based app that runs on both my PC and my Android tablet. I can @import a stylesheet from the CDN, but then it will always use the CDN's version of the font, even though I have the font installed on my PC.

By adding the `local()` calls that tries to find the font on the OS first, I can point to a single CDN url for both my PC and my tablet, and my PC will prefer the locally installed fonts, while my tablet will load the fonts remotely.

This may be unexpected behavior, so I'll leave that up to the community, but it is a small optimization for my use case.